Remove unnecessary abstract parent class for parser
