+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/ConfigMojo.java
@@ -14,7 +14,7 @@ import com.microsoft.azure.maven.webapp.configuration.Deployment;
 import com.microsoft.azure.maven.webapp.configuration.SchemaVersion;
 import com.microsoft.azure.maven.webapp.handlers.WebAppPomHandler;
 import com.microsoft.azure.maven.webapp.models.WebAppOption;
-import com.microsoft.azure.maven.webapp.parser.V2ConfigurationParser;
+import com.microsoft.azure.maven.webapp.parser.ConfigurationParser;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.appservice.AzureAppService;
 import com.microsoft.azure.toolkit.lib.appservice.model.JavaVersion;
@@ -81,7 +81,7 @@ public class ConfigMojo extends AbstractWebAppMojo {
     private static final String PRICE_TIER_NOT_AVAIL = "The price tier \"%s\" is not available for current OS or runtime, use \"%s\" instead.";
     private static final String NO_JAVA_WEB_APPS = "There are no Java Web Apps in current subscription, please follow the following steps to create a new one.";
     private static final String LONG_LOADING_HINT = "It may take a few minutes to load all Java Web Apps, please be patient.";
-    private static final String[] configTypes = { "Application", "Runtime", "DeploymentSlot" };
+    private static final String[] configTypes = {"Application", "Runtime", "DeploymentSlot"};
     private static final String SETTING_DOCKER_IMAGE = "DOCKER_CUSTOM_IMAGE_NAME";
     private static final String SETTING_REGISTRY_SERVER = "DOCKER_REGISTRY_SERVER_URL";
     private static final String SETTING_REGISTRY_USERNAME = "DOCKER_REGISTRY_SERVER_USERNAME";
@@ -274,7 +274,7 @@ public class ConfigMojo extends AbstractWebAppMojo {
 
         final boolean isJBoss = isJBossRuntime(WebContainer.fromString(configuration.getWebContainer()));
         final PricingTier defaultPricingTierFromRuntime = isJBoss ?
-            WebAppConfiguration.DEFAULT_JBOSS_PRICING_TIER : WebAppConfiguration.DEFAULT_PRICINGTIER;
+                WebAppConfiguration.DEFAULT_JBOSS_PRICING_TIER : WebAppConfiguration.DEFAULT_PRICINGTIER;
         final String currentPricingTier = configuration.getPricingTier();
         final List<String> availablePriceList = getAvailablePricingTierList(configuration.getOs(), WebContainer.fromString(configuration.getWebContainer()));
         String defaultPricingTier = currentPricingTier;
@@ -358,7 +358,7 @@ public class ConfigMojo extends AbstractWebAppMojo {
                 defaultPricingTier = defaultPricingTierEnu.getSize();
             }
             final String pricingTier = queryer.assureInputFromUser("pricingTier", defaultPricingTier, availablePriceList,
-                String.format(PRICING_TIER_PROMPT, defaultPricingTier));
+                    String.format(PRICING_TIER_PROMPT, defaultPricingTier));
             builder.pricingTier(pricingTier);
         }
         return builder.build();
@@ -377,7 +377,8 @@ public class ConfigMojo extends AbstractWebAppMojo {
     }
 
     private WebContainer getRuntimeConfigurationForWindowsOrLinux(OperatingSystem os,
-        WebAppConfiguration.WebAppConfigurationBuilder<?, ?> builder, WebAppConfiguration configuration) throws MojoFailureException {
+                                                                  WebAppConfiguration.WebAppConfigurationBuilder<?, ?> builder,
+                                                                  WebAppConfiguration configuration) throws MojoFailureException {
         String defaultJavaVersion = Objects.toString(configuration.getJavaVersionOrDefault());
         final List<String> validJavaVersions = getValidJavaVersions();
         if (!validJavaVersions.contains(defaultJavaVersion)) {
@@ -398,20 +399,20 @@ public class ConfigMojo extends AbstractWebAppMojo {
         }
         String webContainerOrDefault = configuration.getWebContainerOrDefault();
         final List<String> validWebContainers = getAvailableWebContainer(os, javaVersion,
-            Utils.isJarPackagingProject(this.project.getPackaging()));
+                Utils.isJarPackagingProject(this.project.getPackaging()));
         if (!validWebContainers.contains(webContainerOrDefault)) {
             Log.warn(String.format("'%s' is not supported.", webContainerOrDefault));
             webContainerOrDefault = WebAppConfiguration.DEFAULT_CONTAINER.toString();
         }
         final String webContainerInput = queryer.assureInputFromUser(WEB_CONTAINER, webContainerOrDefault,
-            validWebContainers,
-            String.format(COMMON_PROMPT, WEB_CONTAINER, webContainerOrDefault));
+                validWebContainers,
+                String.format(COMMON_PROMPT, WEB_CONTAINER, webContainerOrDefault));
         builder.javaVersion(javaVersion.toString()).webContainer(webContainerInput);
         return WebContainer.fromString(webContainerInput);
     }
 
-    private WebAppConfiguration.WebAppConfigurationBuilder<?, ?> getRuntimeConfigurationOfDocker(WebAppConfiguration.WebAppConfigurationBuilder<?, ?> builder,
-                                                                        WebAppConfiguration configuration) throws MojoFailureException {
+    private WebAppConfiguration.WebAppConfigurationBuilder<?, ?> getRuntimeConfigurationOfDocker(
+            WebAppConfiguration.WebAppConfigurationBuilder<?, ?> builder, WebAppConfiguration configuration) throws MojoFailureException {
         final String image = queryer.assureInputFromUser("image", configuration.image, NOT_EMPTY_REGEX, null, null);
         final String serverId = queryer.assureInputFromUser("serverId", configuration.serverId, null, null, null);
         final String registryUrl = queryer.assureInputFromUser("registryUrl", configuration.registryUrl, null, null,
@@ -463,7 +464,7 @@ public class ConfigMojo extends AbstractWebAppMojo {
     }
 
     private WebAppConfiguration getWebAppConfigurationWithoutValidation() throws AzureExecutionException {
-        return new V2ConfigurationParser(this).getWebAppConfiguration();
+        return new ConfigurationParser(this).getWebAppConfiguration();
     }
 
     private WebAppConfiguration chooseExistingWebappForConfiguration()
